OSIS-4954 move ProgramManager to auth/roles
