+++ assessments/business/score_encoding_list.py
@@ -30,8 +30,9 @@ from decimal import Decimal, Context, Inexact
 from django.db import transaction
 from django.utils.translation import gettext_lazy as _
 
-from base.models import academic_year, session_exam_calendar, exam_enrollment, program_manager, tutor, offer_year, \
+from base.models import academic_year, session_exam_calendar, exam_enrollment, tutor, offer_year, \
     learning_unit_year
+from base.auth.roles import program_manager
 from base.models.enums import exam_enrollment_justification_type
