+++ assessments/views/pgm_manager_administration.py
@@ -42,7 +42,7 @@ from base.auth.roles.entity_manager import EntityManager
 from base.models.offer_type import OfferType
 from base.models.offer_year import OfferYear
 from base.models.person import Person
-from base.models.program_manager import ProgramManager
+from base.auth.roles.program_manager import ProgramManager
 from base.views.mixins import AjaxTemplateMixin
 
 ALL_OPTION_VALUE = "-"
@@ -309,7 +309,7 @@ def _get_programs(academic_yr, entity_list, manager_person, an_offer_type):
 
 def _get_entity_program_managers(entity, academic_yr):
     entities = get_managed_entities(entity)
-    return mdl.program_manager.find_by_management_entity(entities, academic_yr)
+    return base.auth.roles.program_manager.find_by_management_entity(entities, academic_yr)
 
 
 def find_values(key_value, json_repr):
