+++ base/apps.py
@@ -30,8 +30,7 @@ class BaseConfig(AppConfig):
     name = 'base'
 
     def ready(self):
-        from base.models.models_signals import add_to_tutors_group, remove_from_tutor_group, \
-            remove_from_pgm_managers_group, update_person
+        from base.models.models_signals import add_to_tutors_group, remove_from_tutor_group, update_person
         from base.models.utils.lookups import ArrayContainsAny
         from assessments.views.score_encoding import get_json_data_scores_sheets
         # if django.core.exceptions.AppRegistryNotReady: Apps aren't loaded yet.
