+++ base/templatetags/education_group.py
@@ -27,7 +27,7 @@ import waffle
 from django import template
 from django.utils.translation import gettext as _
 
-from base.models import program_manager
+from base.auth.roles import program_manager
 
 register = template.Library()
 
@@ -42,7 +42,7 @@ def have_only_access_to_certificate_aims(user, education_group_year):
     program_manager to be redirected to update_certificate_aims
     """
     return program_manager.is_program_manager(user, education_group=education_group_year.education_group) \
-        and not any((user.is_superuser, user.person.is_faculty_manager, user.person.is_central_manager))
+           and not any((user.is_superuser, user.person.is_faculty_manager, user.person.is_central_manager))
 
 
 @register.simple_tag(takes_context=True)
